Move useCanvasInteraction into layout
