+++ src/content-views/FlamechartView.js
@@ -1,12 +1,11 @@
 // @flow
 
-import type {Interaction, MouseMoveInteraction} from '../useCanvasInteraction';
 import type {
   Flamechart,
   FlamechartStackFrame,
   FlamechartStackLayer,
 } from '../types';
-import type {Rect, Size} from '../layout';
+import type {Interaction, MouseMoveInteraction, Rect, Size} from '../layout';
 
 import {
   ColorView,
