+++ src/layout/index.js
@@ -8,3 +8,4 @@ export * from './VerticalScrollView';
 export * from './View';
 export * from './geometry';
 export * from './layouter';
+export * from './useCanvasInteraction';
